make bench only runnable with pro
