+++ operators/src/pro/mod.rs
@@ -2,5 +2,6 @@
 
 pub mod adapters;
 pub mod cache;
+pub mod engine;
 pub mod meta;
 pub mod ml;
